Api Controller의 Request 내용을 Main Controller로 이전
