+++ remora/src/main/java/remora/remora/Api/ApiController.java
@@ -16,28 +16,19 @@ import java.util.List;
  */
 
 @RestController
-@RequestMapping("/")
 public class ApiController {
 
     private ApiService apiService = new ApiService();
 
-    @GetMapping(value = "/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
-    UploadResponseDto uploadVideo(@RequestParam("originVideo") List<MultipartFile> files,
-                                  @RequestParam("needTranslate") String needTranslate) throws IOException {
-
-        Boolean trans = needTranslate.equals("true");
-
-        UploadRequestDto uploadReqDto = new UploadRequestDto(files, trans);
+    public UploadResponseDto uploadVideo(UploadRequestDto uploadReqDto) throws IOException {
         return apiService.uploadVideo(uploadReqDto);
     }
 
-    @PutMapping("/upload")
-    SimpleResponseDto changeVideo(@RequestBody UploadRequestDto uploadReqDto){
+    public SimpleResponseDto changeVideo(UploadRequestDto uploadReqDto){
         return apiService.changeVideo(uploadReqDto);
     }
 
-    @DeleteMapping("/upload")
-    SimpleResponseDto deleteVideo(@RequestBody DeleteRequestDto deleteReqDto){
+    public SimpleResponseDto deleteVideo(DeleteRequestDto deleteReqDto){
         return apiService.deleteVideo(deleteReqDto);
     }
 }
