+++ remora/src/main/java/remora/remora/MainControl/MainController.java
@@ -1,10 +1,18 @@
 package remora.remora.MainControl;
 
 import java.io.File;
+import java.io.IOException;
+import java.util.List;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.http.MediaType;
+import org.springframework.web.bind.annotation.*;
 
+import org.springframework.web.multipart.MultipartFile;
+import remora.remora.Api.ApiController;
+import remora.remora.Api.dto.DeleteRequestDto;
+import remora.remora.Api.dto.SimpleResponseDto;
+import remora.remora.Api.dto.UploadRequestDto;
+import remora.remora.Api.dto.UploadResponseDto;
 import remora.remora.FrameExtraction.FrameExtractionController;
 import remora.remora.FrameExtraction.dto.FrameExtractionRequestDto;
 import remora.remora.FrameExtraction.dto.FrameExtractionResponseDto;
@@ -15,9 +23,33 @@ import remora.remora.Translation.dto.TranslationResponseDto;
 
 @RestController
 public class MainController {
+    FrameExtractionController frameExtractionController = new FrameExtractionController();
+    TranslationController translationController = new TranslationController();
+    ApiController apiController = new ApiController();
+
+    @GetMapping(value = "/upload", consumes = MediaType.MULTIPART_FORM_DATA_VALUE)
+    UploadResponseDto uploadVideo(@RequestParam("originVideo") List<MultipartFile> files,
+                                  @RequestParam("needTranslate") String needTranslate) throws IOException {
+
+        Boolean trans = needTranslate.equals("true");
+        UploadRequestDto request = new UploadRequestDto(files, trans);
+
+        return apiController.uploadVideo(request);
+    }
+
+    @PutMapping("/upload")
+    SimpleResponseDto changeVideo(@RequestBody UploadRequestDto uploadReqDto){
+        return apiController.changeVideo(uploadReqDto);
+    }
+
+    @DeleteMapping("/upload")
+    SimpleResponseDto deleteVideo(@RequestBody DeleteRequestDto deleteReqDto){
+        return apiController.deleteVideo(deleteReqDto);
+    }
+
     @GetMapping("/test/extraction")
     public void testExtraction() {
-        FrameExtractionController frameExtractionController = new FrameExtractionController();
+        //FrameExtractionController frameExtractionController = new FrameExtractionController();
         FrameExtractionRequestDto request = new FrameExtractionRequestDto();
         FrameExtractionResponseDto response = null;
 
@@ -30,7 +62,7 @@ public class MainController {
 
     @GetMapping("/test/translate")
     public void testTranslate() {
-        TranslationController translationController = new TranslationController();
+        //TranslationController translationController = new TranslationController();
         TranslationRequestDto request = new TranslationRequestDto();
         TranslationResponseDto response = new TranslationResponseDto();
